set dtype to bfloat16

+++ training/pretraining/components/muon_optim.py
@@ -51,7 +51,7 @@ def muon_update(
 
     # Orthonormalize gradient and update momentum in orthonormal space
     grad_orthonorm = newtonschulz5(grad, use_bf16)
-    momentum.lerp_(grad_orthonorm, 1-beta2)
+    momentum.lerp_(grad_orthonorm.to(momentum.dtype), 1-beta2)
 
     if nesterov:
         update = grad_orthonorm.lerp(momentum, 1-beta1)
